--- frontend/Dockerfile
@@ -1,19 +1,0 @@
-# Frontend Dockerfile
-FROM python:3.11-slim
-
-WORKDIR /app
-
-# Copy requirements first for better caching
-COPY frontend/requirements.txt requirements.txt
-
-# Install dependencies
-RUN pip install --no-cache-dir -r requirements.txt
-
-# Copy application code
-COPY . .
-
-# Expose port
-EXPOSE 8501
-
-# Command to run the application
-CMD ["streamlit", "run", "app.py", "--server.address=0.0.0.0", "--server.port=8501"]
