--- frontend/requirements.txt
@@ -1,4 +1,0 @@
-# Frontend requirements
-streamlit>=1.24.0
-httpx>=0.24.1
-python-dotenv>=1.0.0
