+++ opinion_analysis/Dockerfile
@@ -1,4 +1,4 @@
-# Backend Dockerfile
+# opinion_analysis Dockerfile
 FROM python:3.11-slim
 
 WORKDIR /app
@@ -11,7 +11,7 @@ RUN apt-get update && apt-get install -y --no-install-recommends \
     && rm -rf /var/lib/apt/lists/*
 
 # Copy requirements first for better caching
-COPY  backend/requirements.txt requirements.txt
+COPY  requirements.txt requirements.txt
 
 # Install dependencies
 RUN pip install --no-cache-dir -r requirements.txt
@@ -23,7 +23,7 @@ RUN python -m spacy download zh_core_web_sm
 COPY . .
 
 # Expose port
-EXPOSE 8000
+EXPOSE 8501
 
 # Command to run the application
-CMD ["uvicorn", "main:app", "--host", "0.0.0.0", "--port", "8000"]
+CMD ["streamlit", "run", "main.py", "--server.address=0.0.0.0", "--server.port=8501"]
