+++ opinion_analysis/main.py
@@ -1,0 +1,18 @@
+
+import streamlit as st
+
+
+pages = {
+    "prompt 管理": [
+        st.Page(
+            "page/chatbot.py",
+            title="Opinion Analysis Chatbot",
+            icon="📊",
+        ),
+
+    ]
+}
+
+
+pg = st.navigation(pages)
+pg.run()
