+++ opinion_analysis/requirements.txt
@@ -1,12 +1,13 @@
-fastapi>=0.103.1
-uvicorn>=0.23.2
-pydantic>=2.3.0
-langchain>=0.0.267
-langgraph>=0.0.17
+fastapi==0.103.1
+uvicorn==0.23.2
+pydantic==2.3.0
+langchain==0.1.0
+langgraph==0.0.17
+langserve==0.0.30
 openai>=1.0.0
 google-generativeai>=0.3.0
 beautifulsoup4>=4.12.2
-httpx>=0.24.1
+httpx==0.24.1
 python-dotenv>=1.0.0
 newspaper3k>=0.2.8
 lxml[html_clean]
@@ -15,3 +16,6 @@ spacy>=3.7.0
 -f --index-url https://download.pytorch.org/whl/cpu
 torch
 hf_xet
+streamlit>=1.24.0
+httpx>=0.24.1
+python-dotenv>=1.0.0
